змінено тип вхідних даних з integer на float

+++ завд1.py
@@ -5,9 +5,9 @@ def hello():
 
 def read_abc():
     print ('Введіть сторони a,b,c трикутника')
-    a=int(input())
-    b=int(input())
-    c=int(input())
+    a=float(input())
+    b=float(input())
+    c=float(input())
     if a+b>c and b+c>a and a+c>b and a>0 and b>0 and c>0:
         return (a,b,c)
     raise
